requirements.txt
Django==3.2.8
dlib==19.22.1
face-recognition==1.3.0
numpy==1.21.3
Pillow==8.4.0
scikit-learn==1.0
